requirements.txt
flask==3.1.0
openai==1.58.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==21.2.0
flask-cors==4.0.0